load_cells/requirements.txt
pandas==2.2.3
RPi.GPIO==0.7.1
spidev==3.6
scipy==1.10.1
matplotlib==3.7.1